3.1.1 优化当前App版本的获取方式，防止AppInfo获取失败导致获取当前App版本号错误的Bug。

+++ apkupdater/src/main/java/com/kelin/apkUpdater/UpdateInfo.kt
@@ -36,7 +36,7 @@ interface UpdateInfo {
      * @return 返回你要强制更新的版本号，可以返回 null ，如果返回 null 并且 [.isForceUpdate] 返回 true 的话
      * 则表示所有版本全部强制更新。
      */
-    val forceUpdateVersionCodes: IntArray?
+    val forceUpdateVersionCodes: LongArray?
 
     /**
      * 更新标题，例如"更新以下内容"，用于显示在弹窗中。
